sched/fair: Refactor packing eligible test.
The purpose of is_packing_eligible() test is to check if
we should throwaway the active cpu and use the idle CPU.
Refactor this function such a way that, the throw away
CPU (i.e active_cpu) is marked -1. The follow up patch
use this refactoring to add more checks in this function.

Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
[adharmap: made adjust cpus for packing inline]
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
[Render: Minor fixups]
Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I41c107104c56ba5499bfcd9ca4a3aee0a4706212