sched/core: fix preemption disable tracking during context switch to idle.
schedule() calls __schedule() with preemption disabled. The preemption
latency tracking code make a note of this. If the CPU enters idle there
is no up date to the latency tracking code. When the CPU exits idle,
preemption is enabled by the task that is scheduled in. The latency
tracking code thinks that preemption is disabled the whole time. To fix
this, reset the preemption latency tracking time upon context switching
from the idle task. This is only done for the idle task because we still
want to catch the long preemption disable time during the context switch
which may also involves running time of new idle load balance.

Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: Id37aeea53c2110fefc5b541a17410feaf52fcd9e