BACKPORT: sched/fair: reduce no-hz idle balance for energy aware systems.
We are kicking no-hz idle balance when the CPU has two tasks.
It should be done only when the CPU is overutilized on
energy aware systems. This approach seems to be causing more
number of wakeups on HZ=250 systems compared to HZ=100.

Revisit the no-hz idle scheme and only send the IPI if and
when it is necessary.

Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I6692bc314d0326932ccc6d9e230d9d3de64847bd